Norm and Add condition to loop of replace_env for '\''
